schedule的bug修复

+++ src/components/patrolschedule/PatrolScheduleTable.tsx
@@ -33,10 +33,21 @@ const PatrolScheduleTable = () => {
 
   const handleAddSchedule = async () => {
     try {
-      const response = await axios.post("/api/vehicle/patrolschedule/add", {
-        schedule_date: selectedDate,
-        patrol_team_id: selectedTeam,
-      });
+      if (selectedDate) {
+        const localDate = `${selectedDate.getFullYear()}-${(
+          selectedDate.getMonth() + 1
+        )
+          .toString()
+          .padStart(2, "0")}-${selectedDate
+          .getDate()
+          .toString()
+          .padStart(2, "0")}`;
+
+        const response = await axios.post("/api/vehicle/patrolschedule/add", {
+          schedule_date: localDate,
+          patrol_team_id: selectedTeam,
+        });
+      }
       setDialogOpen(false);
       refetch();
     } catch (error) {
@@ -46,9 +57,20 @@ const PatrolScheduleTable = () => {
 
   const handleDeleteSchedule = async () => {
     try {
-      await axios.delete("/api/vehicle/patrolschedule/delete", {
-        data: { ids: [selectedDate] },
-      });
+      if (selectedDate) {
+        const localDate = `${selectedDate.getFullYear()}-${(
+          selectedDate.getMonth() + 1
+        )
+          .toString()
+          .padStart(2, "0")}-${selectedDate
+          .getDate()
+          .toString()
+          .padStart(2, "0")}`;
+
+        await axios.delete("/api/vehicle/patrolschedule/delete", {
+          data: { schedule_date: localDate },
+        });
+      }
       setConfirmDialogOpen(false);
       refetch();
     } catch (error) {
